refactor(modules): move modules into domain folder
