+++ src/shared/domain/modules/enums.js
